tests/wpt/tests/credential-management/fedcm-multi-idp/multiple-gets-after-abort.https.html

<!DOCTYPE html>
<title>Federated Credential Management API multi IDP get after abort test.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script type="module">
  import {
    set_fedcm_cookie,
    set_alt_fedcm_cookie,
    request_options_with_mediation_required,
    alt_request_options_with_mediation_required,
    fedcm_select_account_promise
  } from '../support/fedcm-helper.sub.js';

  let cookies_promise = Promise.all([set_fedcm_cookie(), set_alt_fedcm_cookie()]);

  promise_test(async t => {
    let first_controller = new AbortController();
    let first_test_options = request_options_with_mediation_required();
    first_test_options.signal = first_controller.signal;
    const first_cred = navigator.credentials.get(first_test_options);

    let second_controller = new AbortController();
    let second_test_options = alt_request_options_with_mediation_required();
    second_test_options.signal = second_controller.signal;
    const second_cred = navigator.credentials.get(second_test_options);

    await cookies_promise;
    second_controller.abort();
    await Promise.all([
      promise_rejects_dom(t, 'AbortError', first_cred),
      promise_rejects_dom(t, 'AbortError', second_cred)
    ]);

    const third_cred = navigator.credentials.get(request_options_with_mediation_required());
    const fourth_cred = navigator.credentials.get(alt_request_options_with_mediation_required());

    // Select first account, i.e. from the `third_cred`.
    await fedcm_select_account_promise(t, 0);

    // NetworkError is returned when another IDP is selected.
    await promise_rejects_dom(t, 'NetworkError', fourth_cred);
    const cred = await third_cred;
    assert_equals(cred.token, "token");
  }, "Multiple gets after aborting a multi IDP request should work");
</script>
